Dashboard Panel + MINOR FIX
-  Total Transaksi Hari ini
- Total profit hari ini
- Employee Rank SQL Algorithm

+++ src/popups/EditProduk.java
@@ -91,7 +91,7 @@ public class EditProduk extends javax.swing.JDialog {
             if (ie.getStateChange() == ItemEvent.SELECTED) {
                 if (ie.getItem().equals("-- Tambah Karat --")) {
                     try {
-                        int karatBaru = Integer.valueOf(JOptionPane.showInputDialog("Input karat :"));
+                        double karatBaru = Double.valueOf(JOptionPane.showInputDialog("Input karat :"));
                         if (karatBaru != 0) {
                             addKarat(karatBaru);
                             comboKarat.setModel(getKarat());
